# Remove some unneeded memory allocation