--- a/mesecons_delayer/init.lua
+++ b/mesecons_delayer/init.lua
@@ -108,10 +108,12 @@ local off_state = {
 		receptor =
 		{
 			state = mesecon.state.off,
+			const_node = true,
 			rules = delayer_get_output_rules
 		},
 		effector =
 		{
+			const_node = true,
 			rules = delayer_get_input_rules,
 			action_on = delayer_activate
 		}