--- a/mesecons_extrawires/tjunction.lua
+++ b/mesecons_extrawires/tjunction.lua
@@ -46,6 +46,7 @@ minetest.register_node("mesecons_extrawires:tjunction_on", {
 	mesecons = {conductor =
 	{
 		state = mesecon.state.on,
+		const_node = true,
 		rules = tjunction_get_rules,
 		offstate = "mesecons_extrawires:tjunction_off"
 	}},
@@ -76,6 +77,7 @@ minetest.register_node("mesecons_extrawires:tjunction_off", {
 	mesecons = {conductor =
 	{
 		state = mesecon.state.off,
+		const_node = true,
 		rules = tjunction_get_rules,
 		onstate = "mesecons_extrawires:tjunction_on"
 	}},