--- a/mesecons_gates/init.lua
+++ b/mesecons_gates/init.lua
@@ -99,8 +99,10 @@ local function register_gate(name, inputnumber, assess, recipe, description)
 		groups = {dig_immediate = 2, overheat = 1},
 		mesecons = { receptor = {
 			state = "off",
+			const_node = true,
 			rules = gate_get_output_rules
 		}, effector = {
+			const_node = true,
 			rules = get_inputrules,
 			action_change = update_gate
 		}}
@@ -118,8 +120,10 @@ local function register_gate(name, inputnumber, assess, recipe, description)
 		groups = {dig_immediate = 2, not_in_creative_inventory = 1, overheat = 1},
 		mesecons = { receptor = {
 			state = "on",
+			const_node = true,
 			rules = gate_get_output_rules
 		}, effector = {
+			const_node = true,
 			rules = get_inputrules,
 			action_change = update_gate
 		}}