--- a/mesecons_insulated/init.lua
+++ b/mesecons_insulated/init.lua
@@ -38,6 +38,7 @@ minetest.register_node("mesecons_insulated:insulated_on", {
 	mesecons = {conductor = {
 		state = mesecon.state.on,
 		offstate = "mesecons_insulated:insulated_off",
+		const_node = true,
 		rules = insulated_wire_get_rules
 	}},
 	on_blast = mesecon.on_blastnode,
@@ -74,6 +75,7 @@ minetest.register_node("mesecons_insulated:insulated_off", {
 	mesecons = {conductor = {
 		state = mesecon.state.off,
 		onstate = "mesecons_insulated:insulated_on",
+		const_node = true,
 		rules = insulated_wire_get_rules
 	}},
 	on_blast = mesecon.on_blastnode,