--- a/mesecons_pistons/init.lua
+++ b/mesecons_pistons/init.lua
@@ -281,6 +281,7 @@ minetest.register_node("mesecons_pistons:piston_normal_off", {
 	sounds = mesecon.node_sound.wood,
 	mesecons = {effector={
 		action_on = piston_on,
+		const_node = true,
 		rules = piston_get_rules,
 	}},
 	on_punch = piston_punch,
@@ -311,6 +312,7 @@ minetest.register_node("mesecons_pistons:piston_normal_on", {
 	sounds = mesecon.node_sound.wood,
 	mesecons = {effector={
 		action_off = piston_off,
+		const_node = true,
 		rules = piston_get_rules,
 	}},
 	on_rotate = piston_rotate_on,
@@ -360,6 +362,7 @@ minetest.register_node("mesecons_pistons:piston_sticky_off", {
 	sounds = mesecon.node_sound.wood,
 	mesecons = {effector={
 		action_on = piston_on,
+		const_node = true,
 		rules = piston_get_rules,
 	}},
 	on_punch = piston_punch,
@@ -390,6 +393,7 @@ minetest.register_node("mesecons_pistons:piston_sticky_on", {
 	sounds = mesecon.node_sound.wood,
 	mesecons = {effector={
 		action_off = piston_off,
+		const_node = true,
 		rules = piston_get_rules,
 	}},
 	on_rotate = piston_rotate_on,