--- a/mesecons_receiver/init.lua
+++ b/mesecons_receiver/init.lua
@@ -68,6 +68,7 @@ mesecon.register_node("mesecons_receiver:receiver", {
 	},
 	mesecons = {conductor = {
 		state = mesecon.state.off,
+		const_node = true,
 		rules = receiver_get_rules,
 		onstate = "mesecons_receiver:receiver_on"
 	}}