--- a/mesecons_solarpanel/init.lua
+++ b/mesecons_solarpanel/init.lua
@@ -21,12 +21,14 @@ mesecon.register_node("mesecons_solarpanel:solar_panel", {
 	groups = {dig_immediate = 3},
 	mesecons = {receptor = {
 		state = mesecon.state.off,
+		const_node = true,
 		rules = mesecon.rules.wallmounted_get
 	}}
 },{
 	groups = {dig_immediate = 3, not_in_creative_inventory = 1},
 	mesecons = {receptor = {
 		state = mesecon.state.on,
+		const_node = true,
 		rules = mesecon.rules.wallmounted_get
 	}},
 })