--- a/mesecons_torch/init.lua
+++ b/mesecons_torch/init.lua
@@ -67,6 +67,7 @@ minetest.register_node("mesecons_torch:mesecon_torch_off", {
 	sounds = mesecon.node_sound.default,
 	mesecons = {receptor = {
 		state = mesecon.state.off,
+		const_node = true,
 		rules = torch_get_output_rules
 	}},
 	on_blast = mesecon.on_blastnode,