--- a/mesecons_walllever/init.lua
+++ b/mesecons_walllever/init.lua
@@ -34,6 +34,7 @@ mesecon.register_node("mesecons_walllever:wall_lever", {
 	mesh="jeija_wall_lever_off.obj",
 	on_rotate = mesecon.buttonlike_onrotate,
 	mesecons = {receptor = {
+		const_node = true,
 		rules = mesecon.rules.buttonlike_get,
 		state = mesecon.state.off
 	}},
@@ -48,6 +49,7 @@ mesecon.register_node("mesecons_walllever:wall_lever", {
 	mesh="jeija_wall_lever_on.obj",
 	on_rotate = false,
 	mesecons = {receptor = {
+		const_node = true,
 		rules = mesecon.rules.buttonlike_get,
 		state = mesecon.state.on
 	}},