## Overview

This PR does three things:

- Clean up the basic info getters like `mesecon.get_conductor`.
- Speed up the functions for finding node links, remove unnecessary allocations.
  - As far as I can tell there can only be one linking rule between two nodes, so the functions have been simplified.
  - The inverted function now returns the actual rule, not the inverted rule.
- Remove unnecessary node copies.
  - Components in the modpack with rule getter functions are passed non-copied nodes. External components can get non-copied nodes as well by putting the `const_node` key in their definitions.

## Benchmarks

I used the benchmarks from https://github.com/minetest-mods/mesecons/pull/578. Keep in mind that reducing garbage collection load may also increase the performance of code outside mesecons.

### Long line of diodes

Before: 0.11s
After: 0.10s (likely lower just due to random chance)

### Block of wires

I turned the block on and off 10 times.

Before: 1.7s
After: 0.8s

### 256 byte memory bank

Before: 1.5s
After: 1.3s

## Tests

I've done manual testing and things work. The tests from https://github.com/minetest-mods/mesecons/pull/605 also pass.