@TurkeyMcMac 2022-07-18
Perhaps someone with the right privilege can run a search of ContentDB for mods using `rules_link_rule_all` and `rules_link_rule_all_inverted`. I know lwwires uses them. If they are used by many external mods, I will provide new implementations of them for compatibility with these mods.