@TurkeyMcMac 2022-07-23
I decided to reimplement the functions regardless.