@SmallJoker 2025-02-23
PR needs rebase and adoption.

Keeping it open because it can improve performance a bit.